[CMakeLists.txt] Add BUILD_ARCH option, default to "native"
When cmake is testing if the -march=native switch works, this may fail
with the combination of older compiler and newer build hardware. If this
occurs then the build can fail due to unsupported defaults in the
compiler.

This was experienced when attempting to build the library in a Spack
configuration with an older compiler.

To correct this the BUILD_ARCH option is added to the CMakeLists.txt
file at the root. If the option is not specified on the command line,
then "-march=native" and "-mcpu=native" are the default switches that
are tested.

When cmake is executed with the BUILD_ARCH option like this:

    cmake -DBUILD_ARCH=icelake-server ..

Then cmake will test the "-march=icelake-server" or
"-mcpu=icelake-server" options. This allows build systems such as Spack
to specify the architecture to build for regardless of the architecture
of the build system.