[CMakeLists.txt] Fix BUILD_ARCH option
The CMake option function is intended for ON/OFF operations. When using
"native" as a default, option would set BUILD_ARCH to "OFF" if not used
on the command line. This would cause the build architecture detection
to fail.

In this fix the option function is used properly and, if not used on the
command line, sets a default of "native". Also, if neither -march or
-mcpu work successfully, "native" is used as a fall-back.