[spack/package.py] Add Spack package source
The Spack packaging system requires a `package.py` file in order to
build a package from source. This file:

- Identifies the source origin
- Specifies build and runtime dependencies
- Provides smoke test capabilities based on the project example code

This Spack project file was originally developed for the ECP E4S
project, but should be generic enough for any purpose.

The package has one variant: cuda. This will include extra files that
are needed for the CUDA variant of the bricks library